修复 P0 正确性问题
- hash_pthreads.c: pieces_hashed 改为 _Atomic，消除进度线程数据竞争
  （TSan 复验 0 报告）；进度线程由 stop 标志 + join 优雅退出，
  不再用 PTHREAD_CANCEL_ASYNCHRONOUS + pthread_cancel（避免 stdio
  锁死锁风险）；队列初始化改用指定初始化器
- ftw.c: stat 改为 lstat，不跟随符号链接，杜绝 ELOOP 中止与内容
  重复哈希；新增 test_ftw_skips_symlinks 回归测试
- hash.c/hash_pthreads.c: pieces*SHA_DIGEST_LENGTH 强转 size_t，
  修复 32 位整数回绕导致的堆溢出
- xmake.lua: 测试目标跟随 --pthreads 选项，哈希测试同时覆盖
  多线程实现；test_hash 在 pthread 模式下设置 threads=2

+++ hash.c
@@ -76,7 +76,7 @@ EXPORT unsigned char *make_hash(struct metafile *m)
 
 	/* allocate memory for the hash string
 	   every SHA1 hash is SHA_DIGEST_LENGTH (20) bytes long */
-	hash_string = malloc(m->pieces * SHA_DIGEST_LENGTH);
+	hash_string = malloc((size_t)m->pieces * SHA_DIGEST_LENGTH);
 	/* allocate memory for the read buffer to store 1 piece */
 	read_buf = malloc(m->piece_length);
